+++ src/apis/api.js
@@ -2,16 +2,19 @@ export async function fetchUserData() {
   const response = await fetch(
     "https://bootcamp-api.codeit.kr/api/sample/user"
   );
-  const result = await response.json();
-  // console.log(result);
-  return result;
+  const body = await response.json();
+  // console.log(body);
+  return body;
 }
 
 export async function fetchFolderData() {
   const response = await fetch(
     "https://bootcamp-api.codeit.kr/api/sample/folder"
   );
-  const result = await response.json();
-  // console.log(result);
-  return result;
+  if (!response.ok) {
+    throw new Error("폴더를 불러오는데 실패했습니다.");
+  }
+  const body = await response.json();
+  // console.log(body);
+  return body;
 }
