+++ src/components/Header.js
@@ -3,12 +3,21 @@ import { fetchFolderData } from "../apis/api";
 import "../styles/Header.css";
 
 function Header() {
-  const [folderData, setFolderData] = useState({ folder: { owner: {} } });
+  // const [folderData, setFolderData] = useState({ folder: { owner: {} } });
+  const [folderData, setFolderData] = useState([]);
+  const [loadingError, setLoadingError] = useState(null);
 
   //폴더 데이터 가져오기
   const handleLoadFolder = async () => {
-    const response = await fetchFolderData();
-    setFolderData(response);
+    let folder;
+    try {
+      folder = await fetchFolderData();
+      setLoadingError(null);
+    } catch (e) {
+      setLoadingError(e);
+    } finally {
+    }
+    setFolderData(folder);
   };
 
   //초기데이터 설정
@@ -28,6 +37,7 @@ function Header() {
           @{folderData.folder.owner.name}
         </span>
         <span className="folder-name">{folderData.folder.name}</span>
+        {loadingError?.message && <span>{loadingError.message}</span>}
       </div>
     </header>
   );
