+++ src/utils/calDate.js
@@ -1,0 +1,8 @@
+export function calDate(createdAt) {
+  const date = new Date(createdAt);
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+
+  return `${year}.${month}.${day}`;
+}
